FIX: 문서수발신 domain 수정

+++ src/main/java/kr/or/kmi/mis/api/doc/model/entity/DocDetail.java
@@ -18,7 +18,7 @@ public class DocDetail {
 
     @Id
     @Column(nullable = false)
-    private int draftId;
+    private Long draftId;
 
     @Column(length = 1)
     private String division;
