+++ src/main/java/kr/or/kmi/mis/api/doc/model/entity/DocMaster.java
@@ -17,7 +17,7 @@ public class DocMaster {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int draftId;
+    private Long draftId;
 
     @Column(nullable = false)
     private Timestamp draftDate;
